Refatorar widget mensagens
